// frontend/src/api/generated/models/choreResponse.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Hestia API
 * Internal API for Hestia household inventory management
 * OpenAPI spec version: 1.0.0
 */
import type { ScheduleType } from './scheduleType';

export interface ChoreResponse {
  id: string;
  name: string;
  schedule_type: ScheduleType;
  schedule_value: number;
  /** @nullable */
  assignee?: string | null;
  /** @nullable */
  last_done_at?: string | null;
  next_due_at: string;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}