// frontend/src/api/generated/models/shoppingListSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hestia API
 * Internal API for Hestia household inventory management
 * OpenAPI spec version: 1.0.0
 */

export type ShoppingListSource = typeof ShoppingListSource[keyof typeof ShoppingListSource];


export const ShoppingListSource = {
  manual: 'manual',
  auto: 'auto',
  recipe: 'recipe',
} as const;